backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
litellm==1.52.0
docling==2.10.0
firebase-admin==6.6.0
pydantic>=2.0.0,<3.0.0
python-multipart==0.0.18
feedparser==6.0.11
beautifulsoup4==4.12.3
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
requests==2.32.3
lxml==5.3.0
